feat: Add utility function for moving time
This commit adds a new utility function called "moveTime" to the "move-time.ts" file. The function is responsible for moving the time forward by a specified amount using the Hardhat network provider. It logs a message before and after moving the time. This utility function can be used for testing and development purposes.