feat: Add queueAndExecute function for queuing and executing proposals
This commit adds a new function called "queueAndExecute" to the "queue-and-execute.ts" file. The function is responsible for queuing and executing proposals in the governance system. It interacts with the GovernorContract and Box contracts to queue a proposal, move time and blocks in development chains, and execute the proposal. The function logs the queueing and execution process, as well as the updated value of the Box contract. This function can be used for automating the process of queuing and executing proposals.

Note: This commit message follows the established convention of using the `feat` type for feature additions.